functions/nas_gspread/requirements.txt
google-cloud-storage==1.13.2
gspread==3.1.0
oauth2client==4.1.3
